diamond spinner!
inspired by http://arqex.com/934/4-balls-10-spinners-css3-animations
I've deciced to play around with some css3 aimations. I really like the
simple visual loops he uses so I've decided to make some of my own.

step one is a simple diamond spinner. I want to do an antispin pattern
but that's going to be a lot more complicated, so this is a essentially
a stepping stone.